Challenge2/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
python-multipart>=0.0.5
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=0.19.0
Pillow>=9.0.0
pytesseract>=0.3.8 